// src/models/ITask.ts
export interface ITask {
    id: number;
    // number: string;
    title: string;
    // description: string;
    // date_of_creation: string;
    // time_at_work: string;
    // expiration_date: string;
    // priority: string;
    // attached_files: any;
    // current_status: number;
    // subtasks: [];
    // comments: [];
}